requirements.txt
zarr<3.0.0
tifffile
scikit-image
fsspec
requests
tqdm
python-dotenv
numpy
Pillow